feat(Test File Creation): ALT-T now creates the test file if it doesn't exist.
Changes:

ALT-C now launches the config panel.
The config panel is available through the JSToolbox Tools menu.
The config allows the user to choose the file template to use when creating the file.

Breaking change:

Because it now creates the files, only a single file extension is used in the config.